Move MultiplayerTags class to song.tag package
